Avoid to reset inuse flag for non recycled ccrequests.
Some ccrequests like CkptScanCc, CkptTsCc and ClearTxCc are not recycled
and will be freed after the request has been processed.
But currently, they are treated as recycled requests and will call
CcRequestBase::Free to reset the inuse flag. This will lead to segment
fault if request is freed before CcRequestBase::Free is called.