Using three underscore to separate dbname and table name.
In MariaDB, a table is identified by database name plus table name.
While in Cassandra, we use one keyspace to store all tables of a
single monographDB cluster. As a result we need to use Cassandra's
table name to store both database name and table name of MariaDB.
Since table names in Cassandra must begin with an alpha-numeric
character and can only contain alpha-numeric characters and
underscores. Hence we use three underscore as the delimeter.