New design of schema operations (#80)
This PR implements the new design of schema operations, specifically CREATE TABLE and DROP table. A table's schema is frequently accessed, rarely modified and needs to be strongly consistent and performant. To this end, we model schema data as a special type of in-memory data (in cc maps) replicated in all shards and governed by concurrency control protocols. Each entry in cc maps has a key and a record. For schema cc maps, the cc entry's key is CatalogKey and the record is CatalogRecord. CatalogRecord includes either (1) a binary representation of a schema, or (2) a pointer to the catalog instance at this node (in LocalCcShards). The former is used when instantiating a table catalog instance at this node and creating the corresponding cc map(s). The latter provides fast access to a view of the current and dirty schemas for online tx's. A view of the dirty schema is necessary for certain schema operations, e.g., creating secondary indexes while not blocking online tx's.

For schema cc maps, four more cc requests are defined: AcquireWriteAll and PostWriteAll to acquire and release write locks/intents at all local shards, and RemoteAcquireAll and RemotePostWriteAll to acquire and release write locks/intents at all remote shards (in all other nodes).  They differ from AcquireWrite and PostWrite in that they acquire or release intents/locks on the schema's cc entry in all shards at a node. This is because a table's schema is conceptually replicated in all shards, with each cc entry keeping a pointer to the schema instance in LocalCcShards at the local node.

A schema operation in a tx consists of multiple stages. For example, CREATE TABLE
(1) Acquires write intents on the schema cc entries at all shards, to prevent concurrent modifications on the same schema.
(2) Flushes a prepare log to the log service. After the prepare log is flushed, the schema operation is guaranteed to succeed. 
(3) Installs the dirty schema. From this time forward, all concurrent tx's will see the dirty schema. 
(4) Creates a physical table in the data store.  
(4) Upgrades the write intents to write locks, to prevent any ongoing tx's from accessing this table. 
(5) Flushes the commit log.
(6) Releases write locks and turns the dirty schema to the committed schema.

To support multi-stage operations, the tx state machine TransactionExecution is refactored to use a stack to bookkeep the machine's state. For CREATE TABLE, at the beginning, a create-table op is pushed to the stack. The create-table op internally bookkeeps that it is in the first stage, processes the first async operation (acquire-write-intent), and further pushes acquire-write-intent to the stack. TransactionExecution continuously checks the top of the stack, until the acquire-write-intent returns and is popped out. On processing the current top of the stack, i.e., create-table, the operation moves on to the second stage flush-prepare-log and pushes it to the stack again. The process continues until the create-table op reaches the final stage and pops itself out from the stack.

With the new design, table catalogs and the corresponding cc maps are initialized lazily. In the distributed setting, we need to load the catalog and create cc maps, even if there is no tx at this node. To this end, a new cc request FetchCatalogCc is defined to load catalogs and create cc maps without the presence of tx's.