Replay the records based on table schema timestamp.
Consider we create table T1 and insert tuple A. Then drop table
T1 and then recreate table T1. In this case, tuple A exists in log,
but it should not be replayed after a crash.
This PR use transaction commit_ts as the insert timestamp for
mariadb_tables in Cassandra.