Refactor TxExecution and TxOperation.
1. introduce RunOperation interface which will register operation
and forward the state machine.
2. introduce Process and PostProcess interface to handle all kinds
of the operations.
3. use ProcessTxRequest to make request type clear and easy to
understand.
4. Fix a bug which forgets to set write_entry in AcquireWrite.
5. Remove the meaningless checkpointer assert.
req.ckpt_ts_ <= cce->last_vali_ts_ compares checkpoint ts with last
validation ts, which is meaningless. Checkpoint ts should be smaller
than all the running transaction, but not related to validation.
When compute checkpoint ts, we have already considered this case.
6.Fix a bug of replaying logs with different terms.
A ccnode can be elected as leader multiple times. If the new leader
happens before the last recovery finish, then the ReplayFinishMsg
may come with different terms. This cause a bug. Ignore the finish
message is the term is old and use mutex to protect
recovered_log_groups_.