Fix a bug that ReadOperation finish the TxRequest too early.
Process(ReadOperation) has short path to skip read ccmap if the
read tuple is already in tx_execution's cache. But it finish the
TxRequest too early where the ReadOperation itself is not finished.
This results in the ha_monograph side can issue a new TxRequest,
but the TxProcessor still processes the old ReadOperation.