Fix bugs of schema operations (#274)
* Fix bugs associated with schemas.
1. Remove the clone() method of cc maps. This method is not safe, as the cloned cc map references the same schema object. If the cc maps belong to difference cc node groups, they need to reference the schema of the corresponding cc node group.
2. For cc requests inheriting from TemplateCcMap, rejects the request if it is toward a non-leader cc node, before obtaining the request's target cc map. This prevents the request from initializing and caching a cc map on a non-leader cc node. When the cc node becomes the leader later, the cached cc map's schema may be already out-dated, but there is no way for us to know.
3. For replay cc requests, rejects the requests if the target cc node is not a leader candidate.
4. Introduce DecrementMemory() in CcShard. The function prevents overflow (less than 0) when decrementing memory usage.
5. When PostWriteAll is directed to leaders of two cc node groups in same physical node, the two cc requests should reference their own records. The record of a PostWriteAllCc has two roles: (1) upload a serialized schema imagem, (2) return a pointer to the schema object cached in the tx service.

Add the APIs of creating pk and sk ccm scanners to the catalog factory. Previously, a ccm scanner is created via cc maps, but cc maps may not be available when a cc map scan starts.
When a new key is emplaced into a cc map, resets the new key's schema to the cc map's schema. Since two cc node groups' leaders may fall into the same physcial node, the input key of a cc request may reference the schema of the first cc node group and the request is directed at a second cc node group. Emplacing in such a case cannot reference the input key's schema, but the schema of the second cc node group.