Wait tx_cluster ready during initializing monograph plugin. (#272)
catalog tables will be read during Mysql startup, but in monograph
they are stored in tx_service(kv store) distributedly. Hence we
should wait for the tx_service cluster is ready. Here the ready means
the replay is finished and cc_streams are established.

Also fix a bug: index scan on remote node may fetch catalog on the
fly. It forgets to convert index table name to base table name when
calling FetchCatalog.