Fix workload balance of TxProcessors. (#341)
Previously, we use thread local variable to record tx_run for each
mysql thread. Consider 100 concurrent mysql threads and this may
result in all the 100 txs running on TxprocessorNo1 and then the
next 100 txs running on TxprocessorNo2 and so on.
To make tx workload balance, we replace thread local variable with
atomic variable in TxService.