merge sk ccmap into template ccmap (#358)
* merge sk ccmap into template ccmap

The reason we had a separate class for sk and pk is that they used to
have different key type. For SK we used to store a compound key type
<sk, pk>, but now both sk and pk are packed into a single binary value
, so we can use the same key type. So there's no reason to keep sk ccmap
as it shares the same logic with template ccmap now.

SecondaryKey is currently obselete. Didn't delete it because we might
still need it if we want to do sk compression.

Co-authored-by: lzxddz@qq.com <lzxddz@qq.com>