Set ReadTimestamp when read type change in ReadLocal Request. (#356)
For catalog read, if it cannot find the catalog entry in catalog
ccmap it will check the table_catalogs_ map. If catalog entry
exists, it will convert the ReadInside read request to OutsideNormal
or OutsideDeleted request directly. Note that the ReadTimestamp
of ReadInside request is tx start_ts, which is not correct for the
commit_ts for catalog entry. As a result, we also need to replace
it with the schema version of catalog entry in table_catalogs_.