Range slice (#405)
This PR adds the support of range partitioning of in-memory cc maps. For range partitioning, a range of data items of a table are mapped to a single cc node group, for concurrency control and caching. A range further consists of slices, each spanning dozens of KBs in the data store. The main goals of slices are two folds. First, we bookkeep whether or not a slice of data items in the data store have been fully cached in memory. Such bookkeeping facilitates range scans: if a slice is fully cached, scanning the slice is an in-memory operation. Second, each data store retrieval fetches a slice of data items upon cache misses, to emulate prefetching. Range partitioning and slicing changes the execution path in several respects:

1. For a point read or scan, the tx needs to first uses a key to locate the target range via binary lookups of the local range cc map. The target range determines which cc node group the read/scan request is directed. For point reads, the key is the input key. For scans, the key is either the scan start key or the last key from last scan batch. Upon locating the range, the tx locks the range in the local range cc map, to prevent concurrent splitting of the range, which may move sub-ranges to other cc node groups. The range is unlocked immediately after the point read or the scan toward the range finishes.

2. For a read request toward a cc map, if the target key is not cached, locates the slice that may contain the key and tries to pin the slice in memory. The pin operation may result in a data store read, if the slice is not fully cached. When the slice is pinned, if the key is still missing, we know for sure the key does not exist and return the result as "deleted" to the caller. When a data store read happens, it brings all data items in the data store into memory.

3. For a range scan, we define a new cc request ScanSliceCc to scan data items in a slice. The tx first locates the slice and tries to pin it. The scan is only allowed to proceed, if the slice is fully cached.

4. For writes, the tx locates and locks the range before acquiring write locks on data items. The read locks are released after the tx has uploaded committed values.

A range's slice specification reflects the data distribution in the data store, excluding the data in memory. Since the data store is only updated during checkpointing, a range's slices are only updated then.