Reorder drop table subop (#455)
Move subop that drops table in data store in between commit_log_op_ and
post_all_lock_op_. Right now it is before commit log is flushed, so if
we crash when we are dropping kv table, and the schema recover request
will release the lock on all nodes for a short period of time causing
the deleted kv table being exposed to user query. We want to make sure
that when we are dropping data store table, we always hold a write lock
on all nodes.