Do not load into ccmap if slice is loaded on split at checkpoint (#504)
We used to load the slice into ccmap at GetPostCkptSlice. However
this would contaminate the lru chain and cause potential OOM issue since
we need to force the load no matter how big the data is to unblock checkpoint.
This is unnecessary, we could've just discard the slice once we've calculated
the slice splitting keys.