Fix coredumop on small node memory (#524)
When scan a table, the last key of last scanned slice was kepted to scan
next slice. The memory address of that key was also kepted to re-locate it
quickly. But the key maybe kickout if it was not locked. For example,
under Occ|ReadCommitted scene.

This pr fix it by deduce lock type, instead of acquire a read-intent on
that key under such scene. Because we don't want to breakout the previous
deduce lock type logic.