TableStatistics remove dependence on TableSchema (#533)
In previous version, TableStatistics hold a pointer to TableSchema for below reasons:

1. Deserialize  sample keys received from remote nodes (KeyT::Deserialize() acquire an argument Schema*).\
2. Estimate records in range.
3. Broadcast sample pool with table schema version.
4. Build empty distribution for new table.

The difficulty is how and when to reset the pointer when alter table schema.

This pr try to reduce dependencies on TableSchema to the minimum, and TableStatistics no longer hold pointer to TableSchema. As a replacement, if TableSchema is acquired, pass it as argument instead of holding a pointer.