Range bucket (#582)
Introduce range bucket cc map that map bucket id to node group.
Each range id is hashed to a bucket id with a hash function, then we rely on bucket map to decide
which node group the range belongs to.
We store the bucket record entry address that owns the range in range record so that we can directly
find and lock the bucket record when trying to lock range record. We also stored all of the ranges that
we have loaded into memory in each bucket's bucket info. This will be used when we try to migrate bucket.