Integrate google bigtable (#642)
This PR implements a few interfaces of DataStoreHandler, but not all.

Range partition is supported. Hash partition is not supported.
MVCC is supported. Creating an index(loading by data storage scanner) is not supported. CI is not supported.

CMake option **WITH_DYNAMO_DB** is replaced with **WITH_KV_STORAGE**, whose value can be: CASSANDRA, DYNAMODB, or BIGTABLE, and the default value is CASSANDRA.

Submit an unstable version to prevent conflicts with the main branch.