Update ObjectCcMap Execute ReplayCc. (#801)
* Load record form datastore before replaying log .

The log of MonographRedis saved the series of command not the final result.
So, we must load the checkpoint record from kvstore before replay the pending commands in log.

For hash partition, fetch the record and backfill asynchronously.
For range partition, load the range that the key belongs to like executing `ReadCc`.

* Fix compile error on range partition

* Set `CcEntry::ckpt_ts_` when `ApplyCc` backfill payload.