Multi core scan optimize (#816)
Multi core scan has significant bottleneck on merging scan result from different cores. The merge was done using a heap, on a tx execution state machine forward. Now it is done when a scan slice cc is finished on each core in parallel.
Another problem is that when we have a lot of cores, after data is sharded across all cores, each core might only end up with very few or no tuples. So if we only scan a single slice at a time, we might enter scan slice cc but actually did nothing after a binary search, which is very cost expensive. To avoid this we will scan beyond slice end if the scan cache is not full.