Improve performance of log replay (#859)
Put the replay cc into a random core at first. The cc req will only be forwarded to cores that needs to replay this log msg.
Recycle ReplayLogCc with a CcRequestPool.
Do not wait in every on_received_msg. Only wait on the last msg received from the log group.