Change redis checkpoint: using serialized bytes of payload instead of copying shared_ptr to flush. (#880)
**Change-1**
In MySql, all update operations is replacing old values. So, we let `payload` using `shared_ptr` data type and just copy `shared_ptr` instead of coping record when do checkpoint or scan. When we update a record, we can decide whether to create a new (shared_ptr) object or directly update the tx record of `shared_ptr` pointed to based on `CcEntry::payload_::use_count() == 1`.

For Redis, many update operations are incremental updates rather than replacing existing values. If we execute like above, we must copy the old values of `payload` and update if `shared_ptr::use_count() != 1`. ( `Copy On Write` )
If `payload` data size is large, the `CopyOnWrite` updation is not cost-effective. And, there is no `Scan` operation in Redis.

So, we change the type of `CcEntry::payload_` to `unique_ptr` and directly deserialize the payload to `FlushRecord` for checkpoint. And all update operations are executed on (`CcEntry::payload_`) existing object.

**Change-2**
For redis, we backfilling from ccmap side asynchronously instead of backfilling from server side. (Notice, read kvstore should not block TxProcessor or bthread.)