Replace self-maintained shard memory with mimalloc heap stats  (#890)
Calculating used memory with redis is really hard since it has so many complicated data structures. Use the memory allocation stats from mimalloc as allocated memory for each shard. We use one mimalloc heap for each shard so the allocated memory from mimalloc heap stats is the total memory allocated on this shard. This number will not only increase the heap memory allocated for cc entries and cc maps, but also the txms. This number will be much accurate to the actual memory used by the process.