Skip kv read if tx command overwrites old value (#919)
For commands that overwrites old value in cce, we don't need to read from kv store. We just need to make sure the new commit ts on this cce is larger than the one on kv. Just fake an existing deleted version with current ts and continue the execution.
Also all of the current optimization on DEL command can be applied to the overwrite cmds as well.