Separate cc node (#942)
Decouples tx service and failover mechanism. If FORK_HM_PROCESS is not turned on when compiling, tx service will be started without any failover support. Failover control is now moved to a separate process, host manager. Host manager will be communicating with tx service process with rpc request to notify leader start and leader stop etc. If FORK_HM_PROCESS is on, tx service will fork the host manager process on start up. In the future we might have a standalone node as host manager that will be an authority to tell each node who has become the new leader of node group.