Refactor UpdateStoreSlice (#979)
Now that data sync is done per range, we can simplify the logic in
update store slice.
Refresh ng leader cache immediately on cc res failure. We used to retry
until the last retry before refreshing since UpdateLeader is a sync rpc
call. But now it is changed to an async call so we don't need to wait
until the tx fails before trying to update the leader cache.