invalidate lock term if it is cleaned during data migarte or range split (#1003)
If a cce is kicked out due to data movement, and there's still a lock on it, that means the lock owner has failed, and we should invalid the lock term on that ng. If we receive any cc request with this term, it should be rejected as it might visit cce that is already kicked out.