send bucket cache to new bucket owner during migration (#1022)
Use UploadBatchRequest to send cache to the new bucket owner during migration.
This is done in the flush data operation during data migration, which we additionally scanned the persisted data in memory and send them to the new owner. The result of the UploadBatchRequest will not affect the flush op since it is a best effort performance improvement move.
Added DeserializeObject to TxObject so that we can recreate the object directly from serialized string.