Skip copying key-records and skip sorting whenever possible (#1074)
Introduce three parameters for range scan: is_require_keys, is_require_recs,
and is require_sort.

If is_require_sort is set, the scanner does merge-sort, else the scanner does
concat.

When a key or record can be skipped, serialize it as an empty key or record,
thus the receiver doesn't have to care about skipping them.

Since keys may not transferred to the requester, it is hard for it to remove
trailing tuples. Instead, RemoteScanSliceCache will do that.

The meaning of the cache size of the scanner has been changed. A basic size if
required if none is_require_keys or is_require_recs is set. As a result, each
scan tuple has a basic size of approximately 40 bytes.

This optimization is available when RANGE_PARTITION_ENABLED. The hash partition
will be removed in MonographDB.