key cache (#1109)
Use cuckoo filter as key cache to reduce slice fetch from kv store. Cuckoo filter is a bloom filter that allows deleting keys from the filter, which perfectly suits for our case.
All keys in a range is saved in a key cache. The key cache is build at runtime in memory when the slices in this range is loaded.
When we can't find a key in memory, we verify if it is in the key cache, if the key is not in the key cache we will skip the kv read.