Point read from kv on cache miss (#1160)
In cases specified by requester, we only fetch the record from kv
instead of loading the whole slice on cache miss. This reduces KV
pressure on cases that we do not expect to visit other keys in the same
slice.
Restructured ReadCc to acquire lock first before pinning the
slice/fetching record. This reduces 1 map lookup on cc request resume.
When a cce is first created, it is in Unknown status. It is kept in
unknown status during the kv read. In cases that we know for sure the
key does not exist, the cce is instantly marked as DELETED. (After
verifying KV or cases like hidden pk). A unknown cce is immediately
removed from ccm if there's no lock on it, meaning all cc requests that
might visit this cce are aborted.
If a ReadCc/ApplyCc is aborted due to kv failure, we need to release the
lock acquired on cce so that there's no orphaned lock left.
The key cache holds all keys cached in memory and all keys in KV store.
We add a key to key cache when the record changes from unknown status to
normal/deleted status, and we remove keys from key cache when we kick
out a deleted status cce.