Fix ApplyCc dirty status check condition and dirty object creation. (#1390)
* Fix read command dirty status check, only check dirty status when the dirty status is set by itself which means this txn has acquired write lock before. So that the uncommitted dirty status won't be read by another txn;

* ApplyCc optimization, create dirty object only when dirty status needs to be checked. The problem been fixed is that the dirty object is created whenever the pending command exists even if the cce is accessed by another txn which just acquires NoLock or ReadIntent and only need to read the payload;

* Fix FetchRecord condition and fix the bug that commands that only reads committed payload (under OCC read) wouldn't fetch record in some case.
The case is that for an uncached key, if there is a pending command like SET or DEL which ignores the kv value, the payload status is still unknown. When the SET command is ongoing, if another txn OCC reads the key or in ReadCommitted isolation level, it should FetchRecord from kv and reads the committed key's value.