StoreHandler UpsertTable decouple writetime (#1466)
DataStoreHandler::UpsertTable decouple writetime and schema version.

Currently, commit_ts is used for both write-time and schema version. write-time
is used to support the last-write-win semantic in the replication Cassandra.

We must roll back to the old catalog version to abort a create index
transaction. However, only by using a larger write-time can we overwrite it.