Reduce pinslice operation (#1221)
* Reduce the pinslice operation during add index

When add an index on a non-empty table, we must deal with the problem that an old version key(e.g., the secondary index key generated from the primary key) overwrites a new version key(typically, from a concurrent DML transactions). In the current implementation, will pin slice when upload the encoded secondary index keys, but the a large number of load slice operation will significantly reduce the performance of add index transaction.
If we know that no concurrent DML transaction are writing new version keys, there is no need to load slice to check for potential duplicate keys.

To achieve this, add a flag named `has_dml_since_ddl_` in the `StoreRange` which is `true` default. And, in details:
1. Set the value of `StoreRange::has_dml_since_ddl_`  of the new index as `false` when add index. 

2. When upload encoded secondary index keys, do not pin slice, and for the emplaced ccentry, the value of its `data_store_size_` is the default(INT32_MAX).

3. When do checkpoint for the dirty secondary index table, determining whether duplicate keys need to be checked, and check whether there are keys written by concurrent DML transactions (whose version is greater than the version of this dirty secondary index table). In detail, add two flag for `ScanSliceDeltaSizeCc`. 
        3.1 One named `is_dirty_` whose default is false. For the dirty index, during this scan process, if no keys written by concurrent DML transaction in this range (that is to say, the value of `StoreRange::has_dml_since_ddl_` is false), set the `data_store_size_` of the ccentry as 0 directly if it is INT32_MAX. 
        3.2 Another flag named `has_dml_since_ddl_` which default is false. And if there are keys written by concurrent DML operation (that is to say, the `cce->commit_ts_` larger than the schema ts), set the its value as true. After `ScanSliceDeltaSizeCc` is finished, if the `has_dml_since_ddl_` is true, then set the value of `StoreRange::has_dml_since_ddl_` as true.

4. When split this range, pass the value of the `has_dml_since_ddl_` of this old range to the new splitted ranges.  
---------

Co-authored-by: yangsw <yangsw@chengzhang.com>