Make Cce Payload type a template (#1518)
In order to run multiple compute engine on the same tx service, we need to accommodate range/hash/object in the same tx service.

Removed most of the ON_KEY_OBJECT macro from cc map and cc entry.
Added a template to cc map, cc entry and cc page, indicating if the payload in the cc entry is multi-versioned. For redis, the payload is always updated in place, and for mysql/mongo, we might keep multiple versions of a key in cce.